+++ changelog.md
@@ -8,6 +8,8 @@ Codion Change Log
 - State no longer extends Value, State.is(), toggle() and value() added, ObservableState no longer extends Observable.
 - package is.codion.common.observable renamed observer.
 - Item.value() renamed back to get(), getOrThrow() added.
+### is.codion.common.model
+- FilterModel.RefreshStrategy removed, no longer necessary after recent refresh related selection improvements.
 ### is.codion.swing.common.model
 - NullableToggleButtonModel.ToggleState removed, get() and set() added, related refactoring.
 - FilterListModel.Builder.ItemsStep.items() parameterless overload added.